fix(deps): update cu-cp and cu-up app version to 4.3.14
feat: add close-f1-connection-on-cell-deactivation feature flag

This bump adds a feature flag called close-f1-connection-on-cell-deactivation to make the CU close
the F1-C connection instead of sending a CU configuration Update when the CU wants to deactivate a
cell. This is a workaround for a bug in the Effnet DU that will respond to a CU Configuration Update
that deactivates cells with a CU Configuration Update Acknowledge without actually deactivating the
cells.  The option is disabled by default. It can be enabled using the following configuration:

<?xml version="1.0"?>
<cu-cp-internal xmlns="http://accelleran.com/ns/yang/accelleran-cucp-internal"
xmlns:xc="urn:ietf:params:xml:ns:netconf:base:1.0" xc:operation="replace">
  <feature-flags xc:operation="create">
    <close-f1-connection-on-cell-deactivation>true</close-f1-connection-on-cell-deactivation>
  </feature-flags>
</cu-cp-internal>